@vincentarelbundock 2022-11-17
Oh nice! A 5x improvement in some models is non-trivial, and I didn't know that trick, so thanks a lot.

`R_NOT_CRAN` is an environment variable I use to skip some tests. You can set it in your `bash` profile, or by calling this inside R: `Sys.getenv("R_NOT_CRAN")`