@vincentarelbundock 2022-11-17
Ah no, I see that the 5x is because of `vcov=FALSE`. Still cool. Thanks.